Use dvipdfmx.def verbatim for xetex.def
The XeTeX \XeTeX(pdf|pic)file primitives do not support a full range
of features, in particular adjustment of image interpolation for
bitmaps. At the same time, using \XeTeXpdffile for mesasuring the bbox
will fail with any pagebox spec: the latter may not start at (0,0),
so the data obtained this way has to be used with image inclusion
in a pdfTeX-like way not the current pdf:epdf approach.

This leaves two options:
- Split the approaches in xetex.def, using the dvipdfmx-like
  one for bitmaps and a separate one for PDFs
- Simply go with the dvipdfmx-like route all round

The latter is slightly slower (using extractbb) but avoids
two separate code lines. It is therefore preferred here.