Use dvips specials for colour in dvipdfmx/xetex
For XeTeX using dvips specials keeps things working with PStricks.

For dvipdfmx, it makes most sense to have the same code as for
XeTeX/dvips as this reduces the total number of approaches to keep
track of.

I've left \c@lor@to@ps, etc., in dvipdfmx although I'm not sure if
they are useful there.